+++ src/App.tsx
@@ -11,7 +11,7 @@ import { worker } from './mocks/worker';
 import LikedStorePage from './components/page/likedStorePage';
 import CoinRechargePage from './components/page/coinRechargePage';
 import { store } from './store';
-import ModalContainer from './components/molecules/modalContainer';
+import ModalContainer from './components/modals/modalContainer';
 
 if (process.env.NODE_ENV === 'development') {
   worker.start({
